bot/requirements.txt
aiohttp
aiotg
python-telegram-bot
